+++ frontend/src/templates/MainTemplate/style.css
@@ -3,7 +3,8 @@ a {
 }
 
 .main-template {
-  background-color: #282c34;
+  background: rgb(23,26,33);
+  background:linear-gradient(135deg, rgb(23, 26, 33) 0%, rgba(42,71,94,1) 35%, rgba(27,40,56,1) 100%);
   min-height: 100vh;
   display: flex;
   flex-direction: column;
